[XLA:GPU] Move constant info to module globals
